+++ app/src/main/kotlin/com/simplemobiletools/calculator/activities/MainActivity.kt
@@ -4,7 +4,6 @@ import android.content.Context
 import android.content.Intent
 import android.os.Bundle
 import android.view.Menu
-import android.view.MenuItem
 import android.view.View
 import android.view.WindowManager
 import com.simplemobiletools.calculator.BuildConfig
@@ -13,6 +12,7 @@ import com.simplemobiletools.calculator.databases.CalculatorDatabase
 import com.simplemobiletools.calculator.dialogs.HistoryDialog
 import com.simplemobiletools.calculator.extensions.config
 import com.simplemobiletools.calculator.extensions.updateViewColors
+import com.simplemobiletools.calculator.extensions.updateWidgets
 import com.simplemobiletools.calculator.helpers.*
 import com.simplemobiletools.commons.extensions.*
 import com.simplemobiletools.commons.helpers.LICENSE_AUTOFITTEXTVIEW
@@ -24,36 +24,42 @@ import me.grantland.widget.AutofitHelper
 class MainActivity : SimpleActivity(), Calculator {
     private var storedTextColor = 0
     private var vibrateOnButtonPress = true
+    private var storedUseCommaAsDecimalMark = false
+    private var decimalSeparator = DOT
+    private var groupingSeparator = COMMA
 
-    lateinit var calc: CalculatorImpl
+    private lateinit var calc: CalculatorImpl
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
         appLaunched(BuildConfig.APPLICATION_ID)
+        setupOptionsMenu()
 
         calc = CalculatorImpl(this, applicationContext)
 
-        btn_plus.setOnClickListener { calc.handleOperation(PLUS); checkHaptic(it) }
-        btn_minus.setOnClickListener { calc.handleOperation(MINUS); checkHaptic(it) }
-        btn_multiply.setOnClickListener { calc.handleOperation(MULTIPLY); checkHaptic(it) }
-        btn_divide.setOnClickListener { calc.handleOperation(DIVIDE); checkHaptic(it) }
-        btn_percent.setOnClickListener { calc.handleOperation(PERCENT); checkHaptic(it) }
-        btn_power.setOnClickListener { calc.handleOperation(POWER); checkHaptic(it) }
-        btn_root.setOnClickListener { calc.handleOperation(ROOT); checkHaptic(it) }
+        btn_plus.setOnClickOperation(PLUS)
+        btn_minus.setOnClickOperation(MINUS)
+        btn_multiply.setOnClickOperation(MULTIPLY)
+        btn_divide.setOnClickOperation(DIVIDE)
+        btn_percent.setOnClickOperation(PERCENT)
+        btn_power.setOnClickOperation(POWER)
+        btn_root.setOnClickOperation(ROOT)
+        btn_minus.setOnLongClickListener { calc.turnToNegative() }
 
-        btn_minus.setOnLongClickListener {
-            calc.turnToNegative()
+        btn_clear.setVibratingOnClickListener { calc.handleClear() }
+        btn_clear.setOnLongClickListener {
+            calc.handleReset()
+            true
         }
-
-        btn_clear.setOnClickListener { calc.handleClear(); checkHaptic(it) }
-        btn_clear.setOnLongClickListener { calc.handleReset(); true }
 
         getButtonIds().forEach {
-            it.setOnClickListener { calc.numpadClicked(it.id); checkHaptic(it) }
+            it.setVibratingOnClickListener { view ->
+                calc.numpadClicked(view.id)
+            }
         }
 
-        btn_equals.setOnClickListener { calc.handleEquals(); checkHaptic(it) }
+        btn_equals.setVibratingOnClickListener { calc.handleEquals() }
         formula.setOnLongClickListener { copyToClipboard(false) }
         result.setOnLongClickListener { copyToClipboard(true) }
 
@@ -61,18 +67,25 @@ class MainActivity : SimpleActivity(), Calculator {
         AutofitHelper.create(formula)
         storeStateVariables()
         updateViewColors(calculator_holder, getProperTextColor())
+        setupDecimalSeparator()
         checkWhatsNewDialog()
         checkAppOnSDCard()
     }
 
     override fun onResume() {
         super.onResume()
+        setupToolbar(main_toolbar)
         if (storedTextColor != config.textColor) {
             updateViewColors(calculator_holder, getProperTextColor())
         }
 
         if (config.preventPhoneFromSleeping) {
             window.addFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON)
+        }
+
+        if (storedUseCommaAsDecimalMark != config.useCommaAsDecimalMark) {
+            setupDecimalSeparator()
+            updateWidgets()
         }
 
         vibrateOnButtonPress = config.vibrateOnButtonPress
@@ -98,25 +111,22 @@ class MainActivity : SimpleActivity(), Calculator {
         }
     }
 
-    override fun onCreateOptionsMenu(menu: Menu): Boolean {
-        menuInflater.inflate(R.menu.menu, menu)
-        updateMenuItemColors(menu)
-        return true
-    }
-
-    override fun onOptionsItemSelected(item: MenuItem): Boolean {
-        when (item.itemId) {
-            R.id.history -> showHistory()
-            R.id.settings -> launchSettings()
-            R.id.about -> launchAbout()
-            else -> return super.onOptionsItemSelected(item)
+    private fun setupOptionsMenu() {
+        main_toolbar.setOnMenuItemClickListener { menuItem ->
+            when (menuItem.itemId) {
+                R.id.history -> showHistory()
+                R.id.settings -> launchSettings()
+                R.id.about -> launchAbout()
+                else -> return@setOnMenuItemClickListener false
+            }
+            return@setOnMenuItemClickListener true
         }
-        return true
     }
 
     private fun storeStateVariables() {
         config.apply {
             storedTextColor = textColor
+            storedUseCommaAsDecimalMark = useCommaAsDecimalMark
         }
     }
 
@@ -147,10 +157,13 @@ class MainActivity : SimpleActivity(), Calculator {
         val faqItems = arrayListOf(
             FAQItem(R.string.faq_1_title, R.string.faq_1_text),
             FAQItem(R.string.faq_1_title_commons, R.string.faq_1_text_commons),
-            FAQItem(R.string.faq_4_title_commons, R.string.faq_4_text_commons),
-            FAQItem(R.string.faq_2_title_commons, R.string.faq_2_text_commons),
-            FAQItem(R.string.faq_6_title_commons, R.string.faq_6_text_commons)
+            FAQItem(R.string.faq_4_title_commons, R.string.faq_4_text_commons)
         )
+
+        if (!resources.getBoolean(R.bool.hide_google_relations)) {
+            faqItems.add(FAQItem(R.string.faq_2_title_commons, R.string.faq_2_text_commons))
+            faqItems.add(FAQItem(R.string.faq_6_title_commons, R.string.faq_6_text_commons))
+        }
 
         startAboutActivity(R.string.app_name, licenses, BuildConfig.VERSION_NAME, faqItems, true)
     }
@@ -185,5 +198,31 @@ class MainActivity : SimpleActivity(), Calculator {
 
     override fun showNewFormula(value: String, context: Context) {
         formula.text = value
+    }
+
+    private fun setupDecimalSeparator() {
+        storedUseCommaAsDecimalMark = config.useCommaAsDecimalMark
+        if (storedUseCommaAsDecimalMark) {
+            decimalSeparator = COMMA
+            groupingSeparator = DOT
+        } else {
+            decimalSeparator = DOT
+            groupingSeparator = COMMA
+        }
+        calc.updateSeparators(decimalSeparator, groupingSeparator)
+        btn_decimal.text = decimalSeparator
+    }
+
+    private fun View.setVibratingOnClickListener(callback: (view: View) -> Unit) {
+        setOnClickListener {
+            callback(it)
+            checkHaptic(it)
+        }
+    }
+
+    private fun View.setOnClickOperation(operation: String) {
+        setVibratingOnClickListener {
+            calc.handleOperation(operation)
+        }
     }
 }