+++ app/src/main/kotlin/com/simplemobiletools/calculator/helpers/CalculatorImpl.kt
@@ -3,17 +3,22 @@ package com.simplemobiletools.calculator.helpers
 import android.content.ComponentName
 import android.content.Context
 import android.content.Intent
+import android.text.format.Formatter
 import android.util.Log
-import androidx.core.content.ContextCompat.startActivity
 import com.simplemobiletools.calculator.R
 import com.simplemobiletools.calculator.models.History
 import com.simplemobiletools.commons.extensions.showErrorToast
 import com.simplemobiletools.commons.extensions.toast
 import net.objecthunter.exp4j.ExpressionBuilder
 import java.math.BigDecimal
-import java.util.logging.Handler
+import java.util.*
 
-class CalculatorImpl(calculator: Calculator, private val context: Context) {
+class CalculatorImpl(
+    calculator: Calculator,
+    private val context: Context,
+    private var decimalSeparator: String = DOT,
+    private var groupingSeparator: String = COMMA
+) {
     private var callback: Calculator? = calculator
 
     private var baseValue = 0.0
@@ -24,6 +29,10 @@ class CalculatorImpl(calculator: Calculator, private val context: Context) {
     private val operations = listOf("+", "-", "×", "÷", "^", "%", "√")
     private val operationsRegex = "[-+×÷^%√]".toPattern()
     private val numbersRegex = "[^0-9,.]".toRegex()
+
+    private val formatter = NumberFormatHelper(
+        decimalSeparator = decimalSeparator, groupingSeparator = groupingSeparator
+    )
 
     init {
         showNewResult("0")
@@ -40,21 +49,21 @@ class CalculatorImpl(calculator: Calculator, private val context: Context) {
     }
 
     private fun zeroClicked() {
-        val valueToCheck = inputDisplayedFormula.trimStart('-').replace(",", "")
+        val valueToCheck = inputDisplayedFormula.trimStart('-').removeGroupSeparator()
         val value = valueToCheck.substring(valueToCheck.indexOfAny(operations) + 1)
-        if (value != "0" || value.contains(".")) {
+        if (value != "0" || value.contains(decimalSeparator)) {
             addDigit(0)
         }
     }
 
     private fun decimalClicked() {
-        val valueToCheck = inputDisplayedFormula.trimStart('-').replace(",", "")
+        val valueToCheck = inputDisplayedFormula.trimStart('-').replace(groupingSeparator, "")
         val value = valueToCheck.substring(valueToCheck.indexOfAny(operations) + 1)
-        if (!value.contains(".")) {
+        if (!value.contains(decimalSeparator)) {
             when {
-                value == "0" && !valueToCheck.contains(operationsRegex.toRegex()) -> inputDisplayedFormula = "0."
-                value == "" -> inputDisplayedFormula += "0."
-                else -> inputDisplayedFormula += "."
+                value == "0" && !valueToCheck.contains(operationsRegex.toRegex()) -> inputDisplayedFormula = "0$decimalSeparator"
+                value == "" -> inputDisplayedFormula += "0$decimalSeparator"
+                else -> inputDisplayedFormula += decimalSeparator
             }
         }
 
@@ -65,10 +74,13 @@ class CalculatorImpl(calculator: Calculator, private val context: Context) {
     private fun addThousandsDelimiter() {
         val valuesToCheck = numbersRegex.split(inputDisplayedFormula).filter { it.trim().isNotEmpty() }
         valuesToCheck.forEach {
-            var newString = Formatter.addGroupingSeparators(it)
+            var newString = formatter.addGroupingSeparators(it)
+
             // allow writing numbers like 0.003
-            if (it.contains(".")) {
-                newString = newString.substringBefore(".") + ".${it.substringAfter(".")}"
+            if (it.contains(decimalSeparator)) {
+                val firstPart = newString.substringBefore(decimalSeparator)
+                val lastPart = it.substringAfter(decimalSeparator)
+                newString = "$firstPart$decimalSeparator$lastPart"
             }
 
             inputDisplayedFormula = inputDisplayedFormula.replace(it, newString)
@@ -91,7 +103,7 @@ class CalculatorImpl(calculator: Calculator, private val context: Context) {
         }
 
         val lastChar = inputDisplayedFormula.last().toString()
-        if (lastChar == ".") {
+        if (lastChar == decimalSeparator) {
             inputDisplayedFormula = inputDisplayedFormula.dropLast(1)
         } else if (operations.contains(lastChar)) {
             inputDisplayedFormula = inputDisplayedFormula.dropLast(1)
@@ -142,7 +154,7 @@ class CalculatorImpl(calculator: Calculator, private val context: Context) {
             return false
         }
 
-        if (!inputDisplayedFormula.trimStart('-').any { it.toString() in operations } && inputDisplayedFormula.replace(",", "").toDouble() != 0.0) {
+        if (!inputDisplayedFormula.trimStart('-').any { it.toString() in operations } && inputDisplayedFormula.removeGroupSeparator().toDouble() != 0.0) {
             inputDisplayedFormula = if (inputDisplayedFormula.first() == '-') {
                 inputDisplayedFormula.substring(1)
             } else {
@@ -192,7 +204,7 @@ class CalculatorImpl(calculator: Calculator, private val context: Context) {
     }
 
     private fun getSecondValue(): Double {
-        val valueToCheck = inputDisplayedFormula.trimStart('-').replace(",", "")
+        val valueToCheck = inputDisplayedFormula.trimStart('-').removeGroupSeparator()
         var value = valueToCheck.substring(valueToCheck.indexOfAny(operations) + 1)
         if (value == "") {
             value = "0"
@@ -212,23 +224,28 @@ class CalculatorImpl(calculator: Calculator, private val context: Context) {
         }
 
         if (lastKey != EQUALS) {
-            val valueToCheck = inputDisplayedFormula.trimStart('-').replace(",", "")
+            val valueToCheck = inputDisplayedFormula.trimStart('-').removeGroupSeparator()
             val parts = valueToCheck.split(operationsRegex).filter { it != "" }
             if (parts.isEmpty()) {
                 return
             }
 
-            baseValue = Formatter.stringToDouble(parts.first())
+            baseValue = parts.first().toDouble()
             if (inputDisplayedFormula.startsWith("-")) {
                 baseValue *= -1
             }
 
-            secondValue = parts.getOrNull(1)?.replace(",", "")?.toDouble() ?: secondValue
+            secondValue = parts.getOrNull(1)?.toDouble() ?: secondValue
         }
 
         if (lastOperation != "") {
             val sign = getSign(lastOperation)
-            val expression = "${baseValue.format()}$sign${secondValue.format()}".replace("√", "sqrt").replace("×", "*").replace("÷", "/")
+            val formattedBaseValue = baseValue.format().removeGroupSeparator()
+            val formatterSecondValue = secondValue.format().removeGroupSeparator()
+            val expression = "$formattedBaseValue$sign$formatterSecondValue"
+                .replace("√", "sqrt")
+                .replace("×", "*")
+                .replace("÷", "/")
 
             try {
                 if (sign == "÷" && secondValue == 0.0) {
@@ -236,20 +253,14 @@ class CalculatorImpl(calculator: Calculator, private val context: Context) {
                     return
                 }
 
-                if (sign == "%") {
-                    val second = secondValue / 100f
-                    "${baseValue.format()}*${second.format()}".replace("√", "sqrt").replace("×", "*").replace("÷", "/")
-                }
-
                 // handle percents manually, it doesn't seem to be possible via net.objecthunter:exp4j. "%" is used only for modulo there
                 // handle cases like 10%200 here
                 val result = if (sign == "%") {
-                    val second = secondValue / 100f
-                    ExpressionBuilder("${baseValue.format().replace(",", "")}*${second.format()}").build().evaluate()
+                    val second = (secondValue / 100f).format().removeGroupSeparator()
+                    ExpressionBuilder("$formattedBaseValue*$second").build().evaluate()
                 } else {
                     // avoid Double rounding errors at expressions like 5250,74 + 14,98
                     if (sign == "+" || sign == "-") {
-
                         val first = BigDecimal.valueOf(baseValue)
                         val second = BigDecimal.valueOf(secondValue)
                         val bigDecimalResult = when (sign) {
@@ -258,7 +269,7 @@ class CalculatorImpl(calculator: Calculator, private val context: Context) {
                         }
                         bigDecimalResult.toDouble()
                     } else {
-                        ExpressionBuilder(expression.replace(",", "")).build().evaluate()
+                        ExpressionBuilder(expression).build().evaluate()
                     }
                 }
 
@@ -268,6 +279,7 @@ class CalculatorImpl(calculator: Calculator, private val context: Context) {
                 }
 
                 showNewResult(result.format())
+
                 // build a logic equation
                 // check the result above ^^^ if it is correct then we will send them back to TB,
                 // else we clear the fields -> open a Toast and ask them to try again until they get it right
@@ -276,21 +288,24 @@ class CalculatorImpl(calculator: Calculator, private val context: Context) {
 
                 val valueToCheck = inputDisplayedFormula.trimStart('-').replace(",", "")
                 val parts = valueToCheck.split(operationsRegex).filter { it != "" }
-                baseValue = Formatter.stringToDouble(parts.first())
+                baseValue = parts.first().toDouble()
                 if (inputDisplayedFormula.startsWith("-")) {
                     baseValue *= -1
                 }
 
                 secondValue = parts.getOrNull(1)?.replace(",", "")?.toDouble() ?: secondValue
 
-                if (baseValue.equals(9.0) && secondValue.equals(5.0) && sign.equals("×") && result.equals(45.0)){
+                if (baseValue.equals(9.0) && secondValue.equals(5.0) && sign.equals("×") && result.equals(45.0)) {
 //                    context.toast("You have provided the correct answer!")
 //                    context.toast("Sending you back to the talkback module")
-                    baseValue = result
-                    val newFormula = expression.replace("sqrt", "√").replace("*", "×").replace("/", "÷")
-                    HistoryHelper(context).insertOrUpdateHistoryEntry(History(null, newFormula, result.format(), System.currentTimeMillis()))
-                    inputDisplayedFormula = result.format()
+                    val newFormula = "${baseValue.format()}$sign${secondValue.format()}"
+                    HistoryHelper(context).insertOrUpdateHistoryEntry(
+                        History(id = null, formula = newFormula, result = result.format(), timestamp = System.currentTimeMillis())
+                    )
                     showNewFormula(newFormula)
+
+                    inputDisplayedFormula = result.format()
+                    baseValue = result
 
                     Log.d("test", "back to TB!")
                     try{
@@ -302,13 +317,11 @@ class CalculatorImpl(calculator: Calculator, private val context: Context) {
                     }catch (e:java.lang.Exception){
                         Log.d("test", e.message.toString())
                     }
-                }
-                else{
+                } else {
                     handleReset()
                     resetValues()
                     context.toast("Incorrect! Please try again.")
                 }
-
             } catch (e: Exception) {
                 context.toast(R.string.unknown_error_occurred)
             }
@@ -353,27 +366,22 @@ class CalculatorImpl(calculator: Calculator, private val context: Context) {
         val lastDeletedValue = inputDisplayedFormula.lastOrNull().toString()
 
         var newValue = inputDisplayedFormula.dropLast(1)
-        if (newValue == "") {
+        if (newValue == "" || newValue == "0") {
             newValue = "0"
+            lastKey = CLEAR
         } else {
             if (operations.contains(lastDeletedValue) || lastKey == EQUALS) {
                 lastOperation = ""
             }
             val lastValue = newValue.last().toString()
             lastKey = when {
-                operations.contains(lastValue) -> {
-                    CLEAR
-                }
-                lastValue == "." -> {
-                    DECIMAL
-                }
-                else -> {
-                    DIGIT
-                }
+                operations.contains(lastValue) -> CLEAR
+                lastValue == decimalSeparator -> DECIMAL
+                else -> DIGIT
             }
         }
 
-        newValue = newValue.trimEnd(',')
+        newValue = newValue.trimEnd(groupingSeparator.single())
         inputDisplayedFormula = newValue
         addThousandsDelimiter()
         showNewResult(inputDisplayedFormula)
@@ -435,4 +443,19 @@ class CalculatorImpl(calculator: Calculator, private val context: Context) {
         addThousandsDelimiter()
         showNewResult(inputDisplayedFormula)
     }
+
+    fun updateSeparators(decimalSeparator: String, groupingSeparator: String) {
+        if (this.decimalSeparator != decimalSeparator || this.groupingSeparator != groupingSeparator) {
+            this.decimalSeparator = decimalSeparator
+            this.groupingSeparator = groupingSeparator
+            formatter.decimalSeparator = decimalSeparator
+            formatter.groupingSeparator = groupingSeparator
+            // future: maybe update the formulas with new separators instead of resetting the whole thing
+            handleReset()
+        }
+    }
+
+    private fun Double.format() = formatter.doubleToString(this)
+
+    private fun String.removeGroupSeparator() = formatter.removeGroupingSeparator(this)
 }